♻️ move splash home element to own component
