+++ package.json
@@ -10,6 +10,7 @@
     "@xkeshi/vue-countdown": "^0.6.0",
     "buefy": "^0.6.5",
     "bulma": "^0.7.1",
+    "dayjs": "^1.5.24",
     "jquery": "^3.3.1",
     "mitt": "^1.1.3",
     "normalize.css": "^8.0.0",
