+++ src/views/home/components/splash-hero/components/browser.vue
@@ -233,6 +233,7 @@
     const jQuery = require('jquery');
 
     export default {
+        name: 'ob-browser',
         data(){
             return {
                 posts: [
