+++ src/views/home/components/splash-hero/scripts/typewriter.exec.js
